Update `ons_deaths` table
The ons_deaths table is now a PatientFrame (with one row per patient) so
it is no longer needed to select one event per patient. The ons_deaths
table now includes the earliest registered death for each patient.
Only very few patients have multiple registered deaths so this change is
very unlikely to impact the results of the study.